fix: parse_field() should always keep buffer->start consistent
It is a possible scenario that parse_field() is unable to find
a particular tag or unable to parse the tag even it is found
e.g. the tag has a non-numerical value. Whatever obstacle is
encountered, parse_field() must keep buffer->start value valid
i.e. always pointing to the next "tag=value" pattern.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>