fix: Introduce a new session RX engine
fix_message_recv() has been the only way to get FIX-messages so far.
Such a method not only affects performance (because each time a message
is received a new buffer is allocated) but also tends to be erroneous as
this method is unable to process multiple messages stored in a socket
and leaves them corrupted.

This patch introduces a new fix_session_recv() method which is free of
fix_message_recv() drawbacks. There is a single rx_buffer per
FIX-session which is used as a storage for all FIX-messages ariving
through the socket.  fix_session_recv() is not intended to parse
FIX-messages on the fly i.e.  parse them while reading from the socket.
On the contrary, it reads as big chunk of data as possible and puts it
in rx_buffer.

Implementation notes: currently not a whole buffer is used to store
data.  Essentially, there is a zone of MAX_MESSAGE_SIZE which is not
used at all.  The purpose of such a zone is to avoid dealing with
discontinuity when a FIX-message is split into two parts which are
stored in different places within rx_buffer. An evident flaw of such an
approach is a MAX_MESSAGE_SIZE hole which is actually never used. But
there is a great advantage of not rewriting FIX-messages parsing engine
as it implicitly relies on continuity of data.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>