fix: Process resend requests
If a Resend Request message is received a single Sequence Reset
(Gap Fill mode) is sent back to cover the whole range of missed
messages.

It is quite reasonable to assume that one should use a reliable
transport protocol e.g. TCP/IP. In such a case there is a much
more powerful engine is responsible for data integrity than
simple FIX messages' sequence numbers. Moreover, we are getting
it just for free. Hence, Resend Request indicates that there is
a dramatic error on either side of communication link and FIX
connection should be closed immediately.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>