fix: Track incoming message sequence numbers
If an incoming message's sequence number does not equal to the
expected incoming sequence number then FIX logout is initiated.
This approach is justified for the reliable network transport
protocols e.g. TCP/IP which ensure data integrity.

fix_session_resend_request() is presented for compatibility
purposes and is not used actually. But if the aforementioned
logic if changed, one will initiate Resend Request procedures.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>