fix: Use permanent buffer for message receipt
At the moment, each time the incoming data are unparsed
a new struct fix_message is allocated which is rather
wasteful because of numerous alloc/free operations.

This patch suggests using a permanent per-session pointer
pointing to the most recently unparsed FIX message. If
parsing was successful then this pointer is to be used
for further analysis.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>