FIX: A new testing suite
This commit introduces a new FIX protocol testing suite. The main
idea is to test FIX functionality against the predetermined set of
test cases. Each test case contains of a number of FIX messages for
two parties: FIX server and FIX client respectively. Each message may
be considered as an expected behavior of each party according to the
FIX protocol specification. Hence, any deviation from the expected
behavior could be considered as a violence to protocol specification
and indicates the libtrading implementation error.

Below are details to current FIX testing suite realization:
* Each test case involves two counterparts: the server and the client
* Server is a passive side. Client is an active one. That is to say
the client makes the server respond by sending it FIX messages.
* Each test case should have its own file.
* Each test case file contains of a number of lines.
* Each line represent a single FIX message.
* Lines that start with 'c' ('C') belong to the client.
* Lines that start with 's' ('S') belong to the server.
* One do not have to specify every field within a FIX message.
Only those fields are to be specified that are to be verified.
* The server receives incoming messages and verifies them agains the
expected ones i.e. the messages that have been read from the test case
file.
* The client transmits the messages which have been markes with
'c' ('C') and verifies the messages received from the server against
the 's' ('S') - messages.
* Logon and Logout are performed automatically. So Logon and Logout
messages should not be specified in a test case file. Any attempts to
perform multiple logons will lead to the test case failure.

Consider an example of a test case file:
c35=1\x0134=2\x01
s35=0\x0134=2\x01

This test case will result in the following messages flow:
* Client ---> Server: Logon initiate
* Server ---> Client: Logon accept
* Client ---> Server: Test Request
* Server verifies the incoming message: only tags 35 and 34 are
verified because only these tags were specified in the test case
file (they should be equal to 1 and 2 respectively)
* Server ---> Client: HeartBeat
* Client verifies the incoming message: only tags 35 and 34 are
verified because only these tags were specified in the test case
file (they should be equal to 0 and 2 respectively)
* Client ---> Server: Logout
* Server ---> Client: Logout accept

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>