FIX: test-fix-client a new testing mode is added
A small performance test is added to the test-fix-client.
When this test is invoked a client starts sending a number
of New Single Orders to the server. Once an Execution Report
arrives, a client calculates the round-trip latency.

The first timestamp is made just before the client sends
a New Single Order, the second timestamp is made just after
an Execution Report arrives. So, the round-trip latency is
the difference between the second and first timestamps
expressed in microseconds. From what has been said, it
is clear that this latency is made of four main parts:
encoding a New Order Single by the client, decoding a
FIX message by the server, encoding an Execution Report
by the server and decoding a FIX message by the client.

A new command line option is added to the test-fix-client.
So, if one would like to run a test she just need to specify
the -t option with the number of Execution Reports sent e.g
test-fix-client -t 10000 [hostname] [port] [version] [sender] [target]
would trigger a test with 10000 Orders sent.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>