FIX: Introduce an explicit state of the message being parsed
When the FIX message that is stored in a buffer is being unparse
we should distinguish between two major cases: the message is
garbled and we should skip it when parsing the buffer's content
and the message is just partially delivered and we should wait
until it will be delivered completely.

Until now there was no reliable way to distinguish between these
two cases. The only thing we knew for sure is that if buffer's
size was greater than the maximum possible FIX message size we
shouldn't wait more data to come once we found out that some
part of the buffer contains a garbled field.

This patch suggests that we should explicitly track the FIX
message state while it is being parsed, namely every function
examining the incoming buffer's data explicitly returns the
value associated with the FIX message state.

For example, if the buffer's data are '8=FIX4.4\x019=-1\x01' we
know for sure that body length is corrupted because of the negative
value so we could skip this part of the message. But if the buffer's
data are '8=FIX4.4\x019=1' we should wait the next set of bytes
because the only thing we know is that body length begins with '1'.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>