FAST: Introduction
Two minor changes were made:
* Remove buffer_put(). This functions  is a part of transferring patch.
Not needed in decoding FAST messages.
* Remove FAST_FIELD_SET_STATE macro. This macro remains from previous
versions. At the moment fast messages do not have their states (undefined,
assigned or empty) because is_none field is enough to handle different states.

This patch implements a generic function to decode the incoming data.
In order to support FAST to the full extent a few additional steps
have to be done. At this moment these steps include the following
points:

* Reset support
* Encoding support
* Sequences / Groups support
* Decimal and Unicode support
* Delta operation for strings

Moreover, it would be nice to have a pair of test client and server
as the other part of libtrading have.

This patch is a skeleton and once it is admitted the other steps
should be based on it.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
[ penberg@kernel.org: cleanups ]
Signed-off-by: Pekka Enberg <penberg@kernel.org>