FAST: Explicit dictionary states are introduced
Until now is_none was the proxy of a dictionary state.  While it was
suitable in the most of the cases it was not FAST Specification's exact
replica. For the purpose of being as close to the FAST Spec as possible
we are switching to three dictionary states: undefined, assigned and
empty. This should also make the code clearer for a side reader.

Essentially, the patch just uses the state variable which can be of
three values instead of is_none variable which could be of two values
only.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>