FAST: Do not use states explicitly outside the library
An application which is using libtrading FAST functionality
must not use FAST states explicitly. Moreover, essentially
it should not know about states at all because they are
implementation details which are to be hidden from others.

Instead, every application should use interface provided by
the library. For example, before sending or receiving a
message it should be initialized by fast_message_init().
If a FAST server would not like to transfer an optional
field, it should use field_set_empty() to let the library
know that a particular field is not present in an outcoming
message.

To achieve the aforementioned goal the patch slightly modifies
the testing suite implementation internals. Messages that are
to be sent/received are stored in a container with reference
numbers starting from 1. Zero element is used to store a "pure"
message i.e. a message which is based on a particular template
and which is initialized via fast_message_init. We need such a
message to correctly bind a template with an open fast session.
These manipulations are essential for delta operator.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>