FAST: Fix delta operator problems while receiving a message
The patch was originally created to fix two problems while
a message is being received (Delta operator only):

a) For all numeric data types the delta which is being
transfered is always of type Signed Integer but fast_decode_uint()
now assumes it being Unsigned Integer which obviously lead to bad
consequences

b) It is possible to receive an optional field with delta operator
applied with absent value. In such a case FAST_STATE_EMPTY should
be set. Our code do not allow such a possibility at the moment.
The patch should fix this flaw.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>