FAST: Fix delta operator problems while transferring a message
The patch was originally created to fix two problems while
a message is being transfered (Delta operator only):

a) For all numeric data types the delta which is being
transfered is always of type Signed Integer but fast_encode_uint()
now assumes it being Unsigned Integer which obviously lead to bad
consequences. Fix a function in a way to transfer Signed Integer
if a particular field has Delta operator applied.

b) Delta Operator assumes that delta is always present in a stream.
This means that Delta Operator encoding should always end with
transfer being made. Encoding functions must not return until
something is pushed into the stream. Fix all encoding functions.

Moreover, this patch slightly adjusts the code to make it easier
to read. The idea is that optional and mandatory fields are
handled in a similar way while being encoded except for a few
optional-specific cases. So the code is amended in a way to
get rid of undesirable 'if'-branches i.e. to arrange it in a
plain manner.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>