FAST: clean encoding functions' code
The message that is in an undefined state (before anything
is transferred) must move to another state once the first
message goes out (either to assigned or empty states). It
must not leave in an undefined state.

This patch slightly adjusts the code to make it easier to read.
The idea is that optional and mandatory fields are handled in a
similar way while being encoded except for a few optional-specific
cases. So the code is amended in a way to get rid of undesirable
'if'-branches i.e. to arrange it in a plain manner

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>