FAST: Do not store pmap_bit inside fast field structure
So far there was a one-to-one correspondence between a
particular fast field and PMAP bit number. That is to
say, if a fast field requires a pmap bit then we could
say for sure what bit should be set/checked. This makes
sense in almost all cases except for a single one, namely
when a decimal field has individual operators. In the latter
case, "the presence of the mantissa field and any related
bits in the presence map are dependent on the presence of
the exponent". Hence, as we are going to support decimals
with individual operators in the future, we can no longer
rely on the one-to-one correspondence between a particular
fast field and PMAP bit associated.

This patch assumes that if one has to check/set PMAP bit
the last active bit is used. That's why, pmap_bit is stored
under fast_pmap structure.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>