FAST: hash-table is used to store fields inside a message
So far the only data structure that stores FAST fields inside
a particular FAST message was a simple linear array which makes
other parts of libtrading and external applications either use
it or develop their own ways to manipulate with such an array.
Obviously, the naive way to look through a particular array
affects the performance pretty much.

To make the life of other components easier and to provide a
way of solving the performance problem, the FAST message itself
should provide an interface to quickly manipulate with FAST
fields.

This patch introduces a hash-table which will store FAST fields.
If needed, this data structure will be changed with no difficulties.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>