FAST: Limit order book data structures modification
There are two general ways the limit order book can
be distributed:
 * Price levels distribution. In this case order book
is distributed explicitly. That is order book's dynamics
represents the successive changes applied to a particular
price level i.e. price level modification, price level
removal, price level creation.
 * Orders' flow distribution. In this case order book
is distributed implicitly. That is one receives the
orders' flow and creates a limit order book by processing
orders received.

Libtrading is to support both ways. However, linear arrays
of bids and asks can create some difficulties in the future.
Hence, the current patch presents a few amendments which are
to be applied: instead of using linear arrays hash tables,
binary trees and lists are used.

Whenever a particular level is sought - hash tables are to
be examined. In order to traverse an order book binary trees
or lists are good candidates.

At the moment glib implementation is used but this option
is adjustable.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>