FAST: Apply incremental messages on the fly
We had the following logic implemented whenever
an orderbook is being subscribed (aka late join):
collect incremental messages and wait until snapshot
is arrived. Apply the snapshot and then incremental
messages stored in memory.

Such an implementation has an evident flaw i.e. we
cannot guarantee the number of incremental messages
being stored in memory. It leads to a lot of memory
allocated/reallocated which influence on the library
performance.

This patch allow incremental messages to be applied
without snapshot received. In other words we do not
have to store incremental messages in memory but
rather apply them on the fly.

All we have to know is incremental messages's sequence
number for a particular instrument. Once it is known,
it becomes easy to apply a snapshot message: we should
accept newer changes and discard older ones.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@kernel.org>