FIX tools: Remove test-fix-client and test-fix-server
This patch introduces a single version of FIX client and server
that are based on libtrading API. These are tools/fix/fix_client
and tools/fix/fix_server respectively. Both operate in multiple
modes.

A client operates in three possible modes (this could be changed
in the future):
* Running test scripts (-m 0 or -m script) (default option)
  $ tools/fix/fix_client -m 0 -h ip -p port -f script
* Sending session level messages (-m 1 or -m session)
  $ tools/fix/fix_client -m 1 -h ip -p port -s sender -t target
* Measuring an average latency of executions reports (-m 2 or -m order)
  $ tools/fix/fix_client -m 2 -h ip -p port -s sender -t target -n orders
There are only two mandatory options: FIX server host and port.
All other options are optional. If not specified, sender-comp-id
is set to "BUYSIDE" and target-comp-id is set to "SELLSIDE".

A server operates in two possible modes (this could be changed
in the future):
* Running test scripts (-m 0 or -m script) (default option)
 $ tools/fix/fix_server -m 0 -p port -f script
* Normal mode (-m 1 or -m session)
 $ tools/fix/fix_server -m 1 -p port
There are only one mandatory option, namely port (-p).
All other options are optional. If not specified, sender-comp-id
is set to "SELLSIDE" and target-comp-id is set to "BUYSIDE".

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@iki.fi>