FIX: Integer to string conversion optimizations
The patch introduces an optimized version of integer
to string function representation. The idea is to lookup
through a static array if an integer is relatively small.
In particular this implementation considers the number to
be small if it is less than 100. Such a lookup array also
decreases the number of divisions that are needed to obtain
decimal representation (instead of dividing the number by 10
we have a series of divisions by 100).

We assume that unsigned integers are 4 bytes long.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@iki.fi>