FAST tools: Server and client code unification
FAST server and client code is organized in the same
way as for FIX server and client respectively. The
fact is that both client and server might operate
in several modes. At the moment FAST client and
server can operate in a 'script' mode only but this
functionality is to be extended in the future

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@iki.fi>