FAST tools: Ping-Pong mode introduced
The patch introduces a simple ping-pong mode:
A client sends (pings) a message to server.
A server receives a message from the client,
decodes it and sends (pongs) a message to the
client. Such a scenario allows to estimate
RTT (round-trip-time) of a particular message.
In other words, the time spent by the client
encoding a message, the server decoding it,
then encoding a message a sending it back to
the client and finally the client decoding
the received message.

$ ./tools/fast/fast_server -m pong -n 100 -p 9000 -t ./tools/fast/scripts/PingPong.xml
FAST server is listening to port 9000...

$ ./tools/fast/fast_client -m ping -n 100 -h localhost -p 9000 -t ./tools/fast/scripts/PingPong.xml
Messages sent: 100
Round-trip time: min/avg/max = 40.0/76.6/238.0 μs

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>
Signed-off-by: Pekka Enberg <penberg@iki.fi>