FIX: Request all messages on resend
When the FIX engine detects a incoming message sequence number that is
higher than expected, request all messages on resend by setting EndSeqNo
to zero which represents infinity.

I was not able to find anything conclusive in the FIX specifications but
QuickFIX/J works this way so this is more compatible with other FIX
engines.

Signed-off-by: Pekka Enberg <penberg@iki.fi>