FAST orderbook: set glasks and glbids to NULL explicitly
Once a particular order book is finalized (ob_fini() is called)
we must not refer to internal structures anymore as they are
released. Hence ob_init() is to make sure that internal structures
are initialized before a user is allowed to refer to them.

Otherwise we are exposed to illegal memory access:

ob_fini();
ob_init();

refer to glasks/glbids will lead to SIGSEGV

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>