FIX: move redundant functionality out of scope
Certain functionality could almost never be applied
directly. It is either venue specific or brings so
little value that it is very unlikely to be used at
all. Keeping such functionality as a part of the
library code is very confusing as it shadows another
part which really brings the value and does all the
vital job.

This patch should introduce a series of changes that
are supposed to restructure libtrading FIX code and
make it clear.

The general idea is to divide the whole functionality
into two parts - the core, a higly optimized part
applicable to everybody's needs as is i.e. without
any modification; and the so called general part -
functionality applicable to the general FIX engine
i.e. the one which follows vanilla FIX specification.
One can think of the latter part in two different ways:
either as an API to some ideal (not necessary existent)
FIX engine or as a reference for anyone who wants to
implement their own real application. Following this
idea the patch moves a part of the FIX functionality
under the '/tools' which is considered to be a place
for examples and very often replace documentation
which we lack of at the moment. If we decide that
certain functionality is not a part a core anymore,
we will factor it out as well. We will keep it under
the '/tools' until the next step is done.

The next step is to extend dialect data structure
with operations which trading venues support in real
life e.g. logon, logout, test request etc

We would then provide a basic implementation of those
methods which users are free to rewrite. It would
prevent code duplicate on the one hand and makes it
very clear to users what they need to implement.

Signed-off-by: Marat Stanichenko <mstanichenko@gmail.com>